API

SplitApplyPlot.apply_scale

SplitApplyPlot.draw

SplitApplyPlot.draw!

SplitApplyPlot.iscontinuous

SplitApplyPlot.apply_scale — Method

apply_scale(scale, uniquevalues, value)

Return the value in scale corresponding to the index of value in uniquevalues. Cycle through scale if it has less entries than uniquevalues.

SplitApplyPlot.draw! — Method

draw!(f, fig, data, [by::NamedTuple], select::Mapping)

Group data by columns specified in by, select columns in select, and plot those values on fig using the function f. f can also be an AbstractPlot, such as Scatter or BarPlot. Return the matrix of axes drawn on fig (this may change).

Examples

julia> using SplitApplyPlot, CairoMakie

julia> data = (a=rand(100), b=rand(100), c=rand(100), d=rand(["a", "b"], 100));

julia> fig = Figure()

julia> draw!(Scatter, fig, data, (marker = :d,), mapping(:a, :b, color=:c))

julia> display(fig)

julia> fig = Figure()

julia> mat = draw!(fig, data, (layout_x = :d,), mapping(:a, :b, color=:c)) do ax, m
    plot!(Scatter, ax, m)
    ax.xticklabelrotation[] = π/2
end

julia> hideinnerdecorations!(mat)

julia> display(fig)

SplitApplyPlot.draw — Method

draw(f, args...; kwargs...)

Create fig::Figure with keyword arguments kwargs, then draw on it using draw!, i.e., draw!(f, fig, args...). Return fig.

Examples

julia> using SplitApplyPlot, CairoMakie

julia> data = (a=rand(100), b=rand(100), c=rand(100), d=rand(["a", "b"], 100));

julia> draw(Scatter, data, (marker = :d,), mapping(:a, :b, color=:c))

SplitApplyPlot.iscontinuous — Method

iscontinuous(v::AbstractVector)

Determine whether v should be treated as a continuous or categorical vector.