API

SplitApplyPlot.AxisEntries

SplitApplyPlot.density

SplitApplyPlot.density

SplitApplyPlot.histogram

SplitApplyPlot.histogram

SplitApplyPlot.iscontinuous

SplitApplyPlot.AxisEntries — Type

AxisEntries(axis::Union{Axis, Nothing}, entries::Vector{Entry}, labels, scales)

Define all ingredients to make plots on an axis. Each scale can be either a CategoricalScale (for discrete collections), such as CategoricalScale(["a", "b"], ["red", "blue"]), or a function, such as log10. Other scales may be supported in the future.

SplitApplyPlot.density — Method

density(; extrema, npoints, kernel, bandwidth)

Fit a kernel density estimation of data.

SplitApplyPlot.histogram — Method

histogram(; bins=automatic, wts=automatic, normalization=:none)

Compute a histogram. bins can be an Int to create that number of equal-width bins over the range of values. Alternatively, it can be a sorted iterable of bin edges. The histogram can be normalized by setting normalization. Possible values are:

:pdf: Normalize by sum of weights and bin sizes. Resulting histogram has norm 1 and represents a PDF.

:density: Normalize by bin sizes only. Resulting histogram represents count density of input and does not have norm 1.

:probability: Normalize by sum of weights only. Resulting histogram represents the fraction of probability mass for each bin and does not have norm 1.

:none: Do not normalize.

Weighted data is supported via the keyword wts.

SplitApplyPlot.iscontinuous — Method

iscontinuous(v::AbstractArray)

Determine whether v should be treated as a continuous or categorical vector.