API

SplitApplyPlot.AxisEntries

SplitApplyPlot.density

SplitApplyPlot.density

SplitApplyPlot.expectation

SplitApplyPlot.expectation

SplitApplyPlot.frequency

SplitApplyPlot.frequency

SplitApplyPlot.histogram

SplitApplyPlot.histogram

SplitApplyPlot.iscontinuous

SplitApplyPlot.linear

SplitApplyPlot.linear

SplitApplyPlot.linesfill

SplitApplyPlot.linesfill!

SplitApplyPlot.smooth

SplitApplyPlot.smooth

SplitApplyPlot.AxisEntries — Type

AxisEntries(axis::Union{Axis, Nothing}, entries::Vector{Entry}, labels, scales)

Define all ingredients to make plots on an axis. Each scale can be either a CategoricalScale (for discrete collections), such as CategoricalScale(["a", "b"], ["red", "blue"]), or a function, such as log10. Other scales may be supported in the future.

SplitApplyPlot.density — Method

density(; extrema, npoints, kernel, bandwidth)

Fit a kernel density estimation of data.

SplitApplyPlot.expectation — Method

expectation(args...)

Compute the expected value of the last argument conditioned on the preceding ones.

SplitApplyPlot.frequency — Method

frequency()

Compute a frequency table of the arguments.

SplitApplyPlot.histogram — Method

histogram(; bins=automatic, wts=automatic, normalization=:none)

Compute a histogram. bins can be an Int to create that number of equal-width bins over the range of values. Alternatively, it can be a sorted iterable of bin edges. The histogram can be normalized by setting normalization. Possible values are:

:pdf: Normalize by sum of weights and bin sizes. Resulting histogram has norm 1 and represents a PDF.

:density: Normalize by bin sizes only. Resulting histogram represents count density of input and does not have norm 1.

:probability: Normalize by sum of weights only. Resulting histogram represents the fraction of probability mass for each bin and does not have norm 1.

:none: Do not normalize.

Weighted data is supported via the keyword wts.

Note

Normalizations are computed withing groups. For example, in the case of normalization=:pdf, sum of weights within each group will be equal to 1.

SplitApplyPlot.iscontinuous — Method

iscontinuous(v::AbstractArray)

Determine whether v should be treated as a continuous or categorical vector.

SplitApplyPlot.linear — Method

linear(; interval)

Compute a linear fit of y ~ 1 + x. An optional named mapping wts determines the weights. Use interval to specify what type of interval the shaded band should represent. Valid values of interval are :confidence delimiting the uncertainty of the predicted relationship, and :prediction delimiting estimated bounds for new data points.

SplitApplyPlot.linesfill! — Method

linesfill(xs, ys, lower, upper; kwargs...)

Line plot with a shaded area between lower and upper. If lower and upper are not given, shaded area is between 0 and ys.

Attributes

Available attributes and their defaults for Combined{SplitApplyPlot.linesfill!, T} where T are:

SplitApplyPlot.linesfill — Method

linesfill(xs, ys, lower, upper; kwargs...)

Line plot with a shaded area between lower and upper. If lower and upper are not given, shaded area is between 0 and ys.

Attributes

Available attributes and their defaults for Combined{SplitApplyPlot.linesfill, T} where T are:

  color       :black
  colormap    :viridis
  colorrange  AbstractPlotting.Automatic()
  fillalpha   0.2
  linestyle   "nothing"
  linewidth   1.0

SplitApplyPlot.smooth — Method

smooth(span=0.75, degreee=2)

Fit a loess model. span is the degree of smoothing, typically in [0,1]. Smaller values result in smaller local context in fitting. degree is the polynomial degree used in the loess model.