API

SplitApplyPlot.AxisEntries

SplitApplyPlot.iscontinuous

SplitApplyPlot.AxisEntries — Type

AxisEntries(axis::Union{Axis, Nothing}, entries::Vector{Entry}, labels, scales)

Define all ingredients to make plots on an axis. Each scale can be either a CategoricalScale (for discrete collections), such as CategoricalScale(["a", "b"], ["red", "blue"]), or a function, such as log10. Other scales may be supported in the future.

SplitApplyPlot.iscontinuous — Method

iscontinuous(v::AbstractVector)

Determine whether v should be treated as a continuous or categorical vector.